+++ x/vm/keeper/statedb.go
@@ -55,13 +55,7 @@ func (k *Keeper) GetFastState(ctx sdk.Context, addr common.Address, key common.H
 
 // GetCodeHash loads the code hash from the database for the given contract address.
 func (k *Keeper) GetCodeHash(ctx sdk.Context, addr common.Address) common.Hash {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefixCodeHash)
-	bz := store.Get(addr.Bytes())
-	if len(bz) == 0 {
-		return common.BytesToHash(types.EmptyCodeHash)
-	}
-
-	return common.BytesToHash(bz)
+	return k.AdaptCodeHash(ctx, addr)
 }
 
 // IterateContracts iterates over all smart contract addresses in the EVM keeper and